Move plot_cumulative() and quantile_binning_and_est_errors() into uncertainty_utils
